#62 Added tests to site filters

+++ bigsurvey/webapp/management/commands/reset_db.py
@@ -1,6 +1,7 @@
 from django.core.management.base import CommandError
 from django.core.management.base import BaseCommand
 from main.settings_test import DATABASES
+import os
 
 
 class Command(BaseCommand):
@@ -17,13 +18,12 @@ class Command(BaseCommand):
         engine = DATABASES['default']['ENGINE']
 
         if 'sqlite3' in engine:
-            import os
-            try:
-                print "Unlinking SQLite3 database"
-                os.unlink(DATABASES['default']['NAME'])
-            except OSError:
-                print "Error: %s - %s" % (OSError.errno, OSError.message)
+            if os.path.exists(DATABASES['default']['NAME']):
+                try:
+                    print "Unlinking %s" % DATABASES['default']['NAME']
+                    os.unlink(DATABASES['default']['NAME'])
+                    print "Unlink success"
+                except OSError:
+                    raise CommandError("Cannot unlink %s. Please unlink it manually and retry" % DATABASES['default']['NAME'])
         else:
-            raise CommandError("Only SQLite3 supported for reset")
-
-        print "Unlink success"
+            raise CommandError("Only SQLite3 can be used in RAM-disk")
